Only run node auto switch once per directory switch

+++ terminal/bash/prompt.sh
@@ -4,10 +4,11 @@
 
 # Detect the node version for this project and switch to it
 __detect_node_version() {
-	# If an .nvmrc exists in the current directory, switch to that node version
-	# if it's not already in use
-	if [[ -f .nvmrc && "$(node -v | cut -c2-)" != "$(cat .nvmrc)" ]]; then
+	# If an .nvmrc exists in the current directory (that we just entered),
+	# switch to that node version if it's not already
+	if [[ -f .nvmrc && "$(node -v | cut -c2-)" != "$(cat .nvmrc)" && "$CURRENT_NODE_AUTO_SWITCH_PWD" != "$PWD" ]]; then
 		nvm use 2> /dev/null || nvm install "$(cat .nvmrc)"
+		export CURRENT_NODE_AUTO_SWITCH_PWD="$PWD"
 	fi
 }
